+++ packages/client/src/observable/internal/testUtils.ts
@@ -220,6 +220,7 @@ export function createClientMockHelper(): MockClientHelper {
     clientCacheKey: {} as any,
     requestContext: {},
     logger,
+    asTypeInterfaceOrObjectMapping: {},
   };
   client.fetchMetadata = vitest.fn();