+++ packages/e2e.generated.catchall/src/generatedNoCheck/ontology/objects/Country_1.ts
@@ -114,7 +114,7 @@ export const Country_1 = {
   type: 'object',
   apiName: 'Country_1',
   osdkMetadata: $osdkMetadata,
-  experimentalDoNotUseMetadata: {
+  internalDoNotUseMetadata: {
     rid: 'ri.ontology.main.object-type.0a276176-8d93-489e-93b4-77673de56b9e',
   },
-} satisfies Country_1 & { experimentalDoNotUseMetadata: { rid: string } } as Country_1;
+} satisfies Country_1 & { internalDoNotUseMetadata: { rid: string } } as Country_1;