+++ packages/maker/src/api/defineOntology.ts
@@ -83,6 +83,7 @@ export async function defineOntology(
   body: () => void | Promise<void>,
   outputDir: string | undefined,
   dependencyFile?: string,
+  randomnessKey?: string,
 ): Promise<OntologyIr> {
   namespace = ns;
   dependencies = {};
@@ -119,7 +120,7 @@ export async function defineOntology(
   if (dependencyFile) {
     writeDependencyFile(dependencyFile);
   }
-  return convertOntologyDefinition(ontologyDefinition);
+  return convertOntologyDefinition(ontologyDefinition, randomnessKey);
 }
 
 export function writeStaticObjects(outputDir: string): void {
@@ -225,7 +226,7 @@ export function buildDatasource(
     }
     : undefined;
   return ({
-    rid: "ri.ontology.main.datasource.".concat(apiName),
+    datasourceName: apiName,
     datasource: definition,
     editsConfiguration: {
       onlyAllowPrivilegedEdits: false,