+++ packages/maker/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @osdk/maker
 
+## 0.13.0-beta.14
+
+### Minor Changes
+
+- 9dd5a63: Intermediary links in oac
+- e62408b: Fix Maker Rid Generation
+
+### Patch Changes
+
+- Updated dependencies [b7edfb5]
+  - @osdk/api@2.5.0-beta.14
+
+## 0.13.0-beta.13
+
+### Minor Changes
+
+- d815339: Allow interface props to be excluded from actions
+
+### Patch Changes
+
+- @osdk/api@2.5.0-beta.13
+
 ## 0.13.0-beta.12
 
 ### Minor Changes